release

+++ packages/shared/src/schemas/learning.ts
@@ -6,7 +6,7 @@ export { userTopicProgressSchema, type UserTopicProgress } from "./topic"
 
 // Recent topics query
 export const recentTopicsQuerySchema = z.object({
-  limit: z.coerce.number().min(1).max(50).optional().default(10),
+  limit: z.coerce.number().int().min(1).max(50).optional().default(10),
 })
 
 export type RecentTopicsQuery = z.infer<typeof recentTopicsQuerySchema>
